Updated light images Set a max render distance Added color to light player light is no based on light color fixed byug preventing player from moving based on tile sheet added switch to shaders

+++ src/com/fisherevans/lrk/rpg/Player.java
@@ -123,7 +123,7 @@ public class Player
         return null;
     }
 
-    public int getLightStength()
+    public float getLightStength()
     {
         Equipment light = getEquipment(Equipment.Position.Light);
         if(light == null)
